research-agent/requirements.txt
langgraph>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0
mcp>=1.0.0,<2.0.0
tavily-python>=0.5.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
